pyproject.toml
[project]
name = "workflow-rag"
version = "0.1.0"
description = "Chroma-based RAG for workflow automation rule prompt."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "chromadb>=0.5.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.7.0",
  "rich>=13.7.0",
  "orjson>=3.10.0",
  "tqdm>=4.66.0",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"
